requirements.txt
openai
Ollama
python-dotenv
langchain-openai
langchain-google-genai